--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "opentelemetry-proto"]
-	path = opentelemetry-proto
-	url = https://github.com/open-telemetry/opentelemetry-proto.git
