+++ src/OddDotNet/Dockerfile
@@ -1,0 +1,22 @@
+﻿FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
+USER $APP_UID
+WORKDIR /app
+EXPOSE 4317
+
+FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
+ARG BUILD_CONFIGURATION=Release
+WORKDIR /src
+COPY ["src/OddDotNet/OddDotNet.csproj", "src/OddDotNet/"]
+RUN dotnet restore "src/OddDotNet/OddDotNet.csproj"
+COPY . .
+WORKDIR "/src/src/OddDotNet"
+RUN dotnet build "OddDotNet.csproj" -c $BUILD_CONFIGURATION -o /app/build
+
+FROM build AS publish
+ARG BUILD_CONFIGURATION=Release
+RUN dotnet publish "OddDotNet.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "OddDotNet.dll"]
