--- src/OddDotNet/FilterDelegate.cs
@@ -1,3 +1,0 @@
-namespace OddDotNet;
-
-public delegate bool FilterDelegate<in TContext>(TContext context) where TContext : class;
