+++ src/OddDotNet/OddDotNet.csproj
@@ -1,20 +1,18 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     </PropertyGroup>
 
+    <!-- Internal proto definitions -->
     <ItemGroup>
-      <PackageReference Include="Aspire.Hosting.Testing" Version="8.2.0" />
-      <PackageReference Include="Grpc.AspNetCore" Version="2.65.0" />
-      <PackageReference Include="Grpc.Tools" Version="2.66.0">
-        <PrivateAssets>all</PrivateAssets>
-        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      </PackageReference>
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.1" />
+        <Protobuf Include="Protos\greet.proto" GrpcServices="Server"/>
     </ItemGroup>
+
+    <!-- OpenTelemetry proto definitions -->
     <ItemGroup>
         <Protobuf Include="../../opentelemetry-proto/opentelemetry/proto/common/v1/common.proto" GrpcServices="Server" ProtoRoot="../../opentelemetry-proto/" />
         <Protobuf Include="../../opentelemetry-proto/opentelemetry/proto/resource/v1/resource.proto" GrpcServices="Server" ProtoRoot="../../opentelemetry-proto/" />
@@ -25,4 +23,15 @@
         <Protobuf Include="../../opentelemetry-proto/opentelemetry/proto/collector/metrics/v1/metrics_service.proto" GrpcServices="Server" ProtoRoot="../../opentelemetry-proto/" />
         <Protobuf Include="../../opentelemetry-proto/opentelemetry/proto/collector/trace/v1/trace_service.proto" GrpcServices="Server" ProtoRoot="../../opentelemetry-proto/" />
     </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Grpc.AspNetCore" Version="2.57.0"/>
+    </ItemGroup>
+
+    <ItemGroup>
+      <Content Include="..\..\.dockerignore">
+        <Link>.dockerignore</Link>
+      </Content>
+    </ItemGroup>
+
 </Project>
