--- src/OddDotNet/OddException.cs
@@ -1,17 +1,0 @@
-namespace OddDotNet;
-
-[Serializable]
-public class OddException : Exception
-{
-    public OddException()
-    {
-    }
-
-    public OddException(string message) : base(message)
-    {
-    }
-
-    public OddException(string message, Exception innerException) : base(message, innerException)
-    {
-    }
-}
