+++ src/OddDotNet/Program.cs
@@ -1,0 +1,18 @@
+using OddDotNet.Services;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+builder.Services.AddGrpc();
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+app.MapGrpcService<LogsService>();
+app.MapGrpcService<MetricsService>();
+app.MapGrpcService<TraceService>();
+app.MapGet("/",
+    () =>
+        "Communication with gRPC endpoints must be made through a gRPC client. To learn how to create a client, visit: https://go.microsoft.com/fwlink/?linkid=2086909");
+
+app.Run();
