+++ src/OddDotNet/Services/LogsService.cs
@@ -1,0 +1,12 @@
+using Grpc.Core;
+using OpenTelemetry.Proto.Collector.Logs.V1;
+
+namespace OddDotNet.Services;
+
+public class LogsService : OpenTelemetry.Proto.Collector.Logs.V1.LogsService.LogsServiceBase
+{
+    public override Task<ExportLogsServiceResponse> Export(ExportLogsServiceRequest request, ServerCallContext context)
+    {
+        return base.Export(request, context);
+    }
+}
