+++ tests/OddDotNet.Aspire.AppHost/OddDotNet.Aspire.AppHost.csproj
@@ -14,6 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\OddDotNet\OddDotNet.csproj" />
     <ProjectReference Include="..\OddDotNet.WebApi.One\OddDotNet.WebApi.One.csproj" />
     <ProjectReference Include="..\OddDotNet.WebApi.Two\OddDotNet.WebApi.Two.csproj" />
   </ItemGroup>
